Added additional options
I have added additional options including the ability to hide the handbrake CLI window, and to allow for importing profiles from the standard handbrake application.

The handbrake arguments have also been changed to a variable so that it doesn't have to be set further down in the script.

I have also changed the sonarr episode scan code to be compliant with any new sonarr updates that remove the drone factory folder and automatic drone factory import.