Changed handling of old files
Changed how the script handles files when it is set to not remove old files.
It will now create a log of previously completed items and exclude them from the list of items to be passed to handbrake.

The script will also now copy content to an In Progress folder when it is set to keep old files, as previously these would still be moved, but just not deleted after completion.
It will take longer for the script to start encoding, as copying may take a while, but this is expected due to the the nature of copy vs move.
It will also not remove any files like .txt or .nfo etc.

I have also added an error message for when the source and destination folders are set to the same location.
This is to make sure that the script does not overwrite the original files.